requirements.txt
gensim==3.2.0
numpy==1.14.0
scikit-learn==0.19.1
scipy==0.19.1
tqdm==4.19.6
spacy==2.0.11
PyYAML==3.12

